+++ src/immutable/index.ts
@@ -1,6 +1,10 @@
-/** @module Immutable */
+/**
+ * This module provides support for the immutable content identifier scheme.
+ *
+ * @module Immutable
+ */
 
 export * from './cid.js';
 export * from './hashes.js';
-export * from './operations.js';
-export * from './schema.js';
+export * from './repository.js';
+export * from './scheme.js';