+++ src/mutable/index.ts
@@ -1,0 +1,9 @@
+/**
+ * This module provides support for the mutable content identifier scheme.
+ *
+ * @module Mutable
+ * @experimental
+ */
+
+export * from './repository.js';
+export * from './scheme.js';